requirements.txt
flask
neo4j
langgraph
faiss-cpu
numpy
prometheus-client
